Add patches for PR#79
PR#79 claims to resolve deadlocks when running igprof. Without this fix,
igprof can be near useless, as it can be impossible to get it to complete
a long-running process. However, the admins are yet to merge this to master
and it's not clear when they will do so. Therefore, I pulled the patches
out of git and plugged them into the patches directory for application
at build time.

https://github.com/igprof/igprof/pull/79